# frontend/openapi.snapshot.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Backup Orchestrator",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/agents/{server_id}/heartbeat": {
            "post": {
                "tags": [
                    "agents"
                ],
                "summary": "Agent Heartbeat",
                "operationId": "agent_heartbeat_api_agents__server_id__heartbeat_post",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "server_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Server Id"
                        }
                    },
                    {
                        "name": "X-Agent-Key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Agent-Key"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AgentHeartbeatRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgentHeartbeatResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/agents/{server_id}/jobs": {
            "get": {
                "tags": [
                    "agents"
                ],
                "summary": "List Agent Jobs",
                "description": "Full-snapshot poll of a server's enabled BackupJobs, for the C#/.NET\nagent to discover what it should be backing up. No destination-path\nfield here -- the agent decides the remote path itself and reports it\nback via POST /api/backup-records (BackupRecord.remote_path).",
                "operationId": "list_agent_jobs_api_agents__server_id__jobs_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "server_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Server Id"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 500,
                            "minimum": 1,
                            "default": 200,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 0,
                            "title": "Offset"
                        }
                    },
                    {
                        "name": "X-Agent-Key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Agent-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedResponse_BackupJobRead_"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/agents/{server_id}/monitoring-config": {
            "get": {
                "tags": [
                    "agents"
                ],
                "summary": "Get Agent Monitoring Config",
                "description": "Resolves which Windows service names this server's agent should\ncheck on each heartbeat: the per-server `Server.monitored_service_names`\noverride if set (including an explicit empty list -- \"monitor\nnothing\"), else the global `Settings.DEFAULT_MONITORED_SERVICE_NAMES`\ndefault. Deliberately `is not None`, never a truthiness check -- an\nexplicit `[]` override must not silently fall through to the global\ndefault.",
                "operationId": "get_agent_monitoring_config_api_agents__server_id__monitoring_config_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "server_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Server Id"
                        }
                    },
                    {
                        "name": "X-Agent-Key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Agent-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgentMonitoringConfigResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/agents/{server_id}/connection-config": {
            "get": {
                "tags": [
                    "agents"
                ],
                "summary": "Get Agent Connection Config",
                "description": "Decrypts and returns this server's FTP/SFTP credentials for the\nagent to consume. Gated by `require_connection_config_key` -- a\nSEPARATE, more restricted secret than the general X-Agent-Key (see\nthat dependency's docstring). Every call is audit-logged, for every\noutcome including denials -- see app/core/audit.py.\n\n`principal` mirrors `require_admin_or_agent_key`'s convention: the\nauthenticated admin `User` if a JWT was used, else `None` (key-header\npath has no User principal) -- used here only to fill in\n`auth_method`/`admin_username` on the audit log row, since the\ndependency itself already logged the \"unauthorized\" outcome (if any)\nbefore this handler ever started running.",
                "operationId": "get_agent_connection_config_api_agents__server_id__connection_config_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "server_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Server Id"
                        }
                    },
                    {
                        "name": "X-Connection-Config-Key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Connection-Config-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgentConnectionConfigResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AgentHeartbeatRequest": {
                "properties": {
                    "reachable": {
                        "type": "boolean",
                        "title": "Reachable"
                    },
                    "disks": {
                        "items": {
                            "$ref": "#/components/schemas/AgentDiskUsageItem"
                        },
                        "type": "array",
                        "title": "Disks"
                    },
                    "metrics": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/AgentMetricsItem"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "services": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/AgentServiceStatusItem"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Services"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "reachable"
                ],
                "title": "AgentHeartbeatRequest"
            },
            "AgentHeartbeatResponse": {
                "properties": {
                    "server": {
                        "$ref": "#/components/schemas/ServerRead"
                    },
                    "disks": {
                        "items": {
                            "$ref": "#/components/schemas/DiskRead"
                        },
                        "type": "array",
                        "title": "Disks"
                    },
                    "alerts_raised": {
                        "items": {
                            "$ref": "#/components/schemas/AlertRead"
                        },
                        "type": "array",
                        "title": "Alerts Raised"
                    },
                    "alerts_resolved": {
                        "items": {
                            "$ref": "#/components/schemas/AlertRead"
                        },
                        "type": "array",
                        "title": "Alerts Resolved"
                    }
                },
                "type": "object",
                "required": [
                    "server",
                    "disks",
                    "alerts_raised",
                    "alerts_resolved"
                ],
                "title": "AgentHeartbeatResponse"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}